+++ application/endpoints/__init__.py
@@ -1,3 +1,3 @@
 import application.endpoints.home
-import application.endpoints.room
+import application.endpoints.session_room
 import application.endpoints.socketio
